fix: resolve object count loading in warmup modal after SOLR sections merge
- Fixed loadObjectStats method to properly use settings store
- Added proper store mapping for settings store access
- Object stats now load correctly before opening warmup modal
- Resolves 'Unable to load object count' error in warmup dialog
